backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
sentence-transformers==2.7.0
numpy>=1.24.0,<2.0.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
huggingface-hub>=0.20.0
torch>=2.0.0
transformers>=4.35.0
# Database drivers (install based on your database choice)
psycopg2-binary>=2.9.9  # PostgreSQL driver (for production)
# For MySQL: mysqlclient or pymysql
# For SQLite: included in Python standard library


